requirements.txt
tensorflow==2.15.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
Pillow==10.3.0
scikit-learn==1.4.2
split-folders==0.5.1
streamlit==1.35.0